services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pyjwt[crypto]==2.9.0
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
bcrypt==4.0.1
